backend/requirements.txt
# Django & REST Framework
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.8.0

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Production server
gunicorn==23.0.0
whitenoise==6.9.0

# Environment variables
python-dotenv==1.2.1

# AI / OpenAI
openai==2.8.1

# Utilities
pillow==11.3.0
requests==2.32.5
beautifulsoup4==4.14.3
httpx==0.28.1
pydantic==2.12.5
PyJWT==2.11.0
